Improve performance of perform_rebinding_with_section
This function was using a full strlen check just to check that the

length was greater than 1 so it could index into position 1. I noticed

that this was pretty slow on iPhone 4, so wrote a little function that

just checks that the string is longer than the given value, then

returns.